cleaned the files, changed versions of  dependencies
