+++ .travis.yml
@@ -3,27 +3,26 @@ jdk:
   - openjdk11
 env:
   global:
-    - HUGO_VERSION=0.58.3
+    - HUGO_VERSION=0.64.1
 install:
   - wget "https://github.com/gohugoio/hugo/releases/download/v${HUGO_VERSION}/hugo_${HUGO_VERSION}_Linux-64bit.deb"
   - sudo dpkg -i "hugo_${HUGO_VERSION}_Linux-64bit.deb"
   - nvm install 8
 script:
   # build main site assets: js, css, etc.
-  - cd ./app && yarn run build
+  - cd ./app && npm run build
   # run service scripts: load medium content, fetch github stars and build docs
-  - cd ../scripts && yarn run build
   - cd ../ && hugo
 branches:
   only:
-    - master
-deploy:
-  provider: pages
-  skip_cleanup: true
-  github_token: $GITHUB_TOKEN
-  on:
-    branch: master
-  local_dir: public
-  target_branch: gh-pages
+    - agrest_1.1
+#deploy:
+#  provider: pages
+#  skip_cleanup: true
+#  github_token: $GITHUB_TOKEN
+#  on:
+#    branch: agrest_1.1
+#  local_dir: public
+#  target_branch: gh-pages
 cache:
-  - yar
+  - npm
