+++ app/package.json
@@ -4,25 +4,25 @@
   "main": "main.js",
   "license": "Apache-2.0",
   "devDependencies": {
-    "autoprefixer": "7.1.2",
-    "css-loader": "0.28.4",
+    "autoprefixer": "9.7.4",
+    "css-loader": "3.4.2",
     "exports-loader": "^0.7.0",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "0.11.2",
-    "html-webpack-plugin": "2.29.0",
+    "extract-text-webpack-plugin": "^4.0.0-beta.0",
+    "file-loader": "5.1.0",
+    "html-webpack-plugin": "3.2.0",
     "node-sass": "^4.13.1",
-    "postcss-loader": "2.0.6",
-    "rimraf": "2.6.1",
-    "sass": "1.20.1",
-    "sass-loader": "^7.3.1",
-    "style-loader": "0.18.2",
-    "url-loader": "0.5.9",
-    "webpack": "3.11.0"
+    "postcss-loader": "3.0.0",
+    "rimraf": "3.0.2",
+    "sass": "1.26.2",
+    "sass-loader": "^8.0.2",
+    "style-loader": "1.1.3",
+    "url-loader": "3.0.0",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11"
   },
   "dependencies": {
-    "add": "^2.0.6",
-    "bootstrap": "4.1.3",
-    "highlight.js": "9.16.2",
+    "bootstrap": "4.4.1",
+    "highlight.js": "9.18.1",
     "jquery": "3.4.1",
     "popper.js": "^1.16.0"
   },
