--- app/styles/mixins/_custom-font-face.scss
@@ -1,16 +1,0 @@
-// CUSTOM FONT-FACE
-@mixin customFontFace($family,$src,$weight: normal,$style: normal) {
-    @font-face {
-        font-family: $family;
-        src: url('#{$src}.eot'); // IE9 compat
-        src: url('#{$src}.eot?#iefix') format('embedded-opentype'), // IE8 and below
-                // url("#{$src}.otf") format("opentype"),
-                url('#{$src}.woff2') format('woff2'), // standards
-                url('#{$src}.woff') format('woff'), // standards
-        url('#{$src}.ttf') format('truetype'), // Safari, Android, iOS
-        url('#{$src}.svg##{$family}') format('svg'); // legacy iOS
-
-        font-style: $style;
-        font-weight: $weight;
-    }
-}
