--- app/styles/scroll/_docbook.scss
@@ -1,402 +1,0 @@
-// DOCBOOK
-// Vars (global)
-@import "~bootstrap/scss/functions";
-
-// DOCS NAV (CONTENTS)
-.nav {
-  padding-left: 0;
-  margin-bottom: 0;
-  list-style: none;
-
-  > li {
-    position: relative;
-    display: block;
-
-    > a {
-      position: relative;
-      display: block;
-      padding: 10px 15px;
-    }
-  }
-}
-
-.docbook {
-  // NAV PADDING
-  padding-bottom: 5%;
-
-  @include media-breakpoint-up(lg) {
-    padding-bottom: map-get($spacers, 5);
-  }
-
-  .doc-version {
-    display: block;
-    text-align: center;
-    text-decoration: none;
-
-    .dropdown-toggle:after {
-      font-size: .875em;
-    }
-
-    > a {
-      font-size: 1.56rem;
-    }
-
-    a {
-      font-weight: 300;
-      text-decoration: none;
-      color: #777;
-
-      &:hover {
-        text-decoration: none;
-      }
-
-      &:active {
-        background-color: inherit;
-      }
-
-      &:focus {
-        outline-style: none;
-      }
-    }
-
-    a.active {
-      background-color: rgba(0, 0, 0, .05);
-    }
-  }
-
-  .toc {
-    display: none;
-  }
-
-  .docs-nav {
-    @include media-breakpoint-up(lg) {
-      position: sticky;
-      max-height: calc(100vh - .5rem);
-    }
-
-    overflow-y: auto;
-    overflow-x: hidden;
-    left: 0;
-    top: .5rem;
-
-    .nav {
-      li {
-        a {
-          padding: 5px 10px;
-          text-decoration: none;
-          border-radius: .3rem;
-
-          @include hover-focus-active {
-            background-color: rgba(0, 0, 0, .05);
-          }
-        }
-      }
-    }
-
-    .nav-link {
-      &.active {
-        background-color: rgba(0, 0, 0, .05);
-      }
-    }
-  }
-
-  // main layout
-  @media screen and (max-width: 1023px) {
-
-    .docs-wrapper {
-      flex-direction: column;
-    }
-
-    .docs-main {
-      order: 2;
-      width: 100%;
-    }
-
-    .docs-side {
-      order: 1;
-      width: 100%;
-      margin: 0;
-
-      .toc-title {
-        display: block;
-      }
-
-      .back-to-top {
-        display: none;
-      }
-    }
-  }
-
-  // links
-  a {
-    text-decoration: underline;
-
-    &:hover {
-      text-decoration: none;
-    }
-
-    &.anchorjs-link {
-      text-decoration: none;
-    }
-  }
-
-  // hr
-  .titlepage hr {
-    margin-bottom: 46px;
-  }
-
-  // headings
-  .toc-title {
-    display: none;
-  }
-
-  h1 {
-    text-align: center;
-  }
-
-  h1.title,
-  .h1.title,
-  > h1,
-  > .h1 {
-    margin-top: map-get($spacers, 6);
-    margin-bottom: map-get($spacers, 3);
-  }
-
-  h3.title {
-    font-size: 20px;
-    font-weight: normal;
-    margin-bottom: 10px;
-    margin-top: 15px;
-  }
-
-  h1, h2, h3, h4, h5, h6,
-  .h1, .h2, .h3, .h4, .h5, .h6 {
-    > a[name] {
-      display: block;
-    }
-  }
-
-  // part
-  .part {
-    // part title(page)
-    .titlepage {
-
-      h1 {
-        font-size: 35px;
-      }
-
-      h2 {
-        font-size: 24px;
-      }
-    }
-  }
-
-  h2 {
-    font-size: 1.7rem;
-  }
-
-  h3 {
-    font-size: 1.4rem;
-    margin-bottom: 1rem;
-    margin-top: 3rem;
-  }
-
-  h4 {
-    font-size: 1.1rem;
-    margin-bottom: 1rem;
-    margin-top: 2rem;
-  }
-
-  // chapters
-  .chapter {
-    padding-bottom: 14px; // FIXED HEADER ANCHOR LINKS OFFSET
-  }
-
-  // section
-  .section {
-    // section title(page)
-    .titlepage h2 {
-      font-size: 18px;
-      font-weight: normal;
-      margin-top: .9em;
-      color: $black;
-    }
-  }
-
-  // lists
-  .media-list {
-
-    .media {
-      margin-top: 30px;
-      padding-bottom: 26px;
-
-      &:first-child {
-        margin-top: 30px;
-      }
-    }
-
-    .media:hover img {
-      filter: none;
-    }
-
-    .media-left {
-      padding-right: 30px;
-
-      img {
-        filter: grayscale(100%) opacity(70%);
-        @include addAnimation(all);
-      }
-    }
-
-    .media-heading {
-      text-decoration: none;
-
-      &:hover {
-        text-decoration: underline;
-      }
-
-      .code {
-        display: block;
-        font-size: 18px;
-        line-height: 20px;
-
-        color: $black;
-      }
-    }
-
-    .media-body p {
-      color: $gray-mid-text;
-    }
-
-    .media-right {
-      min-width: 200px;
-
-      .list-vertbars {
-        margin-bottom: 0;
-
-        a,
-        a:hover {
-          color: $black;
-        }
-      }
-    }
-
-    // rwd
-    @media screen and (min-width: $screen-sm) {
-      .media-body {
-        padding-right: 30px;
-      }
-    }
-
-    @media screen and (min-width: $screen-md) {
-      .media-body {
-        padding-left: 30px;
-      }
-    }
-
-    @media screen and (max-width: $screen-xs-max) {
-      &.itemizedlist {
-        padding-left: 0;
-      }
-
-      .media {
-        margin-top: 25px;
-      }
-
-      .media {
-        /* Prevents distributing space between rows */
-        -ms-align-content: flex-start;
-        -webkit-align-content: flex-start;
-        align-content: flex-start;
-        align-items: center;
-
-        /* IE10 uses display: flexbox */
-        display: -ms-flexbox;
-        -ms-flex-flow: row wrap;
-
-        display: -webkit-flex;
-        -webkit-flex-flow: row wrap;
-      }
-
-      .media-left,
-      .media-body,
-      .media-right {
-        display: inline-block
-      }
-
-      .media-left {
-        width: 20%;
-
-        img {
-          max-width: 100%;
-        }
-      }
-      .media-body {
-        width: 80%;
-
-        p {
-          margin-bottom: 0;
-        }
-      }
-      .media-right {
-        width: 80%;
-        margin-left: 20%;
-        padding-left: 0;
-      }
-    }
-
-    @media screen and (max-width: 480px) {
-      .media-left {
-        width: 30%;
-        padding-right: 20px;
-      }
-      .media-body {
-        width: 70%;
-      }
-      .media-right {
-        width: 70%;
-        margin-left: 30%;
-      }
-    }
-  }
-
-  // section
-  .section + .section {
-    margin-top: 20px;
-    margin-bottom: 40px;
-  }
-
-  // tables
-  .pure-table {
-    margin-bottom: 10px;
-    border-color: #e4e4e4;
-
-    thead {
-      background-color: #f1f1f1;
-    }
-
-    th,
-    td {
-      border-color: #e4e4e4;
-      font-size: 0.9em;
-    }
-
-    code {
-      color: #4e4e4e;
-    }
-  }
-
-  caption {
-    font-size: 16px;
-    padding: 25px 0 15px;
-    text-align: left;
-    font-style: normal;
-    color: inherit;
-  }
-
-  th {
-    font-weight: normal;
-    text-transform: uppercase;
-    font-size: 12px;
-  }
-}
