+++ content/page/docs/overview.html
@@ -9,7 +9,7 @@ docsMenuTitle: "Overview"
  <h2 id="what-is-agrest"><a class="anchor" href="#what-is-agrest"></a>1. What is Agrest</h2> 
  <div class="sectionbody"> 
   <div class="paragraph"> 
-   <p>Agrest is a set of tools to build web APIs for your data. You would use it to expose a data store (such as a relational database) as a RESTful web service. To achieve that goal Agrest defines a simple HTTP/JSON-based protocol (described in <a href="protocol.html">Agrest Protocol</a>) and a server-side framework (described in <a href="framework.html">Agrest Framework</a>). The protocol ensures that any client, be it JavaScript, a mobile app or maybe a server-side app, can make sense of any Agrest-enabled service. The framework is written in Java and helps you to build services adhering to the protocol.</p> 
+   <p>Agrest is a set of tools to build web APIs for your data. You would use it to expose a data store (such as a relational database) as a RESTful web service. To achieve that goal Agrest defines a simple HTTP/JSON-based protocol (described in <a href="/protocol">Agrest Protocol</a>) and a server-side framework (described in <a href="/framework">Agrest Framework</a>). The protocol ensures that any client, be it JavaScript, a mobile app or maybe a server-side app, can make sense of any Agrest-enabled service. The framework is written in Java and helps you to build services adhering to the protocol.</p> 
   </div> 
   <div class="paragraph"> 
    <p>Agrest is open source and is distributed under the terms of <a href="https://raw.githubusercontent.com/agrestio/agrest/master/LICENSE.txt">Apache 2.0 license</a></p>
