+++ content/page/docs/workflow.html
@@ -310,7 +310,7 @@ public class CategoryResource {
    </div> 
    <div class="imageblock text-center"> 
     <div class="content"> 
-     <img src="/images/bookstore_ide_prjct.png" alt="bookstore ide prjct"> 
+     <img src="/images/workflow/bookstore_ide_prjct.png" alt="bookstore ide prjct"> 
     </div> 
    </div> 
    <div class="sect3"> 
@@ -570,7 +570,7 @@ components:
    <div class="sect3"> 
     <h4 id="2-attach-agrest-protocol-definition"><a class="anchor" href="#2-attach-agrest-protocol-definition"></a>2.3.2. 2. Attach Agrest protocol definition</h4> 
     <div class="paragraph"> 
-     <p>The Agrest protocol file <a href="#protocol">protocol.yaml</a> contains definition of all <a href="protocol.html#control-parameters">Control Parameters</a>. Just place this <code>protocol.yaml</code> in the catalog were your main <code>bookstore-api.yaml</code> file is located (e.g. 'src/main/java/resources').</p> 
+     <p>The Agrest protocol file <a href="#protocol">protocol.yaml</a> contains definition of all <a href="/protocol#control-parameters">Control Parameters</a>. Just place this <code>protocol.yaml</code> in the catalog were your main <code>bookstore-api.yaml</code> file is located (e.g. 'src/main/java/resources').</p> 
     </div> 
    </div> 
    <div class="sect3">
