+++ layouts/page/docs-index.html
@@ -20,7 +20,7 @@
 
                 </div>
 
-                <div class="col-12 col-lg-8 col-xl-9 py-3 cd-content">
+                <div class="col-12 col-lg-8 col-xl-9 py-3 pl-lg-2 cd-content">
                     <article>
                         {{ .Content }}
                     </article>
