+++ layouts/partials/navbar-docs.html
@@ -1,5 +1,4 @@
 <div >
-
     <nav class="navbar navbar-expand-md navbar-light">
         <div class="container">
             <a class="navbar-brand" href="https://agrest.io/">
